# backend/docs/swagger/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: GameLink 平台 API，包含健康检查、认证与管理端能力。
  title: GameLink API
  version: 0.3.0
paths:
...
  /admin/stats/audit/overview:
    get:
      parameters:
      - description: 开始时间
        in: query
        name: date_from
        type: string
      - description: 结束时间
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 审计总览（按实体/动作汇总）
      tags:
      - Admin/Stats
  /admin/stats/audit/trend:
    get:
      parameters:
      - description: 开始时间
        in: query
        name: date_from
        type: string
      - description: 结束时间
        in: query
        name: date_to
        type: string
      - description: 实体类型
        enum:
        - order
        - payment
        - player
        - game
        - review
        - user
        in: query
        name: entity
        type: string
      - description: 动作
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 审计趋势（日）
      tags:
      - Admin/Stats
  /admin/stats/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Dashboard 概览
      tags:
      - Admin/Stats
  /admin/stats/orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 订单状态统计
      tags:
      - Admin/Stats
  /admin/stats/revenue-trend:
    get:
      parameters:
      - description: 天数，默认7
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 收入趋势（日）
      tags:
      - Admin/Stats
  /admin/stats/top-players:
    get:
      parameters:
      - description: 数量，默认10
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Top 陪玩师排行
      tags:
      - Admin/Stats
  /admin/stats/user-growth:
    get:
      parameters:
      - description: 天数，默认7
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 用户增长（日）
      tags:
      - Admin/Stats
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"